course-management-backend-microservice/course_management/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
django-redis==6.0.0
python-dotenv==1.1.0
celery==5.5.3
django-celery-results==2.6.0
redis==6.2.0
